Number can be integer and float

+++ hassium/builtins/bool.c
@@ -23,7 +23,7 @@ bool obj_is_true(struct obj *obj, struct vm *vm) {
   if (obj == &true_obj) return true;
   if (obj == &false_obj) return false;
   if (obj->type == OBJ_NUM) {
-    float val = obj_num_val(obj);
+    int val = obj_num_val(obj);
     return val != 0;
   }
   return obj->type != OBJ_NONE;
